requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
langgraph==0.2.39
langchain-core==0.3.12
python-dotenv==1.0.1

faiss-cpu
sentence-transformers
numpy